+++ graphile/graphile-many-to-many/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.14](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.13...graphile-many-to-many@1.2.14) (2026-01-11)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.13](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.12...graphile-many-to-many@1.2.13) (2026-01-10)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.12](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.11...graphile-many-to-many@1.2.12) (2026-01-09)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.11](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.10...graphile-many-to-many@1.2.11) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.10](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.9...graphile-many-to-many@1.2.10) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.9](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.8...graphile-many-to-many@1.2.9) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.8](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.7...graphile-many-to-many@1.2.8) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.7](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.6...graphile-many-to-many@1.2.7) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.6](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.5...graphile-many-to-many@1.2.6) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.5](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.4...graphile-many-to-many@1.2.5) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.4](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.3...graphile-many-to-many@1.2.4) (2026-01-08)
+
+**Note:** Version bump only for package graphile-many-to-many
+
 ## [1.2.3](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.2...graphile-many-to-many@1.2.3) (2026-01-07)
 
 **Note:** Version bump only for package graphile-many-to-many