+++ graphile/graphile-plugin-connection-filter/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.14](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.13...graphile-plugin-connection-filter@2.6.14) (2026-01-11)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.13](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.12...graphile-plugin-connection-filter@2.6.13) (2026-01-10)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.12](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.11...graphile-plugin-connection-filter@2.6.12) (2026-01-09)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.11](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.10...graphile-plugin-connection-filter@2.6.11) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.10](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.9...graphile-plugin-connection-filter@2.6.10) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.9](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.8...graphile-plugin-connection-filter@2.6.9) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.8](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.7...graphile-plugin-connection-filter@2.6.8) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.7](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.6...graphile-plugin-connection-filter@2.6.7) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.6](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.5...graphile-plugin-connection-filter@2.6.6) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.5](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.4...graphile-plugin-connection-filter@2.6.5) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
+## [2.6.4](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.3...graphile-plugin-connection-filter@2.6.4) (2026-01-08)
+
+**Note:** Version bump only for package graphile-plugin-connection-filter
+
 ## [2.6.3](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter@2.6.2...graphile-plugin-connection-filter@2.6.3) (2026-01-07)
 
 **Note:** Version bump only for package graphile-plugin-connection-filter