+++ graphql/codegen/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.23.3](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@2.23.2...@constructive-io/graphql-codegen@2.23.3) (2026-01-09)
+
+### Bug Fixes
+
+- react query mode output should be default for generate ([#593](https://github.com/constructive-io/constructive/issues/593)) ([e63f839](https://github.com/constructive-io/constructive/commit/e63f8397b479daf2081e43ef9d160da66de794f0))
+
 ## [2.23.2](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-codegen@2.23.1...@constructive-io/graphql-codegen@2.23.2) (2026-01-07)
 
 **Note:** Version bump only for package @constructive-io/graphql-codegen