+++ jobs/knative-job-service/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.13](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.12...@constructive-io/knative-job-service@0.6.13) (2026-01-10)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.12](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.11...@constructive-io/knative-job-service@0.6.12) (2026-01-09)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.11](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.10...@constructive-io/knative-job-service@0.6.11) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.10](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.9...@constructive-io/knative-job-service@0.6.10) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.9](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.8...@constructive-io/knative-job-service@0.6.9) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.8](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.7...@constructive-io/knative-job-service@0.6.8) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.7](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.6...@constructive-io/knative-job-service@0.6.7) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.6](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.5...@constructive-io/knative-job-service@0.6.6) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.5](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.4...@constructive-io/knative-job-service@0.6.5) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.4](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.3...@constructive-io/knative-job-service@0.6.4) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
 ## [0.6.3](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.2...@constructive-io/knative-job-service@0.6.3) (2026-01-07)
 
 **Note:** Version bump only for package @constructive-io/knative-job-service