+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.21](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.20...@constructive-io/cli@5.6.21) (2026-01-11)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.20](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.19...@constructive-io/cli@5.6.20) (2026-01-10)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.19](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.18...@constructive-io/cli@5.6.19) (2026-01-10)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.18](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.17...@constructive-io/cli@5.6.18) (2026-01-10)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.17](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.16...@constructive-io/cli@5.6.17) (2026-01-09)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.16](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.15...@constructive-io/cli@5.6.16) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.15](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.14...@constructive-io/cli@5.6.15) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.14](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.13...@constructive-io/cli@5.6.14) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.13](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.12...@constructive-io/cli@5.6.13) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.12](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.11...@constructive-io/cli@5.6.12) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.11](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.10...@constructive-io/cli@5.6.11) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.10](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.9...@constructive-io/cli@5.6.10) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.6.9](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.8...@constructive-io/cli@5.6.9) (2026-01-08)
+
+**Note:** Version bump only for package @constructive-io/cli
+
 ## [5.6.8](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.6.7...@constructive-io/cli@5.6.8) (2026-01-07)
 
 **Note:** Version bump only for package @constructive-io/cli