+++ postgres/pgsql-seed/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.12](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.11...pgsql-seed@0.4.12) (2026-01-09)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.11](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.10...pgsql-seed@0.4.11) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.10](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.9...pgsql-seed@0.4.10) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.9](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.8...pgsql-seed@0.4.9) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.8](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.7...pgsql-seed@0.4.8) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.7](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.6...pgsql-seed@0.4.7) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.6](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.5...pgsql-seed@0.4.6) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.5](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.4...pgsql-seed@0.4.5) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
+## [0.4.4](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.3...pgsql-seed@0.4.4) (2026-01-08)
+
+**Note:** Version bump only for package pgsql-seed
+
 ## [0.4.3](https://github.com/constructive-io/constructive/compare/pgsql-seed@0.4.2...pgsql-seed@0.4.3) (2026-01-07)
 
 **Note:** Version bump only for package pgsql-seed